requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pdfplumber==0.11.4
nltk==3.9.1
pydantic==2.6.4
httpx==0.27.0
openai==1.30.1
jinja2==3.1.4
python-multipart==0.0.9